@@ -12,59 +12,58 @@
     "generate-types": "gql-gen --config codegen.yml",
     "dev:ssr": "ng run reference-storefront:serve-ssr",
     "serve:ssr": "node dist/server/main.js",
-    "build:ssr": "ng build --prod && ng run reference-storefront:server:production",
+    "build:ssr": "ng build --configuration=production && ng run reference-storefront:server:production",
     "build:ssr:docker": "ng build --configuration=production,docker && ng run reference-storefront:server:docker",
     "prerender": "ng run reference-storefront:prerender"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.9",
-    "@angular/cdk": "^13.3.7",
-    "@angular/common": "^13.3.9",
-    "@angular/compiler": "^13.3.9",
-    "@angular/core": "^13.3.9",
-    "@angular/forms": "^13.3.9",
-    "@angular/platform-browser": "^13.3.9",
-    "@angular/platform-browser-dynamic": "^13.3.9",
-    "@angular/platform-server": "^13.3.9",
-    "@angular/router": "^13.3.9",
-    "@angular/service-worker": "^13.3.9",
-    "@apollo/client": "^3.6.5",
-    "@fortawesome/angular-fontawesome": "^0.7.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.31",
-    "@fortawesome/free-brands-svg-icons": "^5.15.0",
-    "@fortawesome/free-solid-svg-icons": "^5.15.0",
-    "@nguniversal/express-engine": "^13.1.1",
-    "apollo-angular": "^3.0.1",
-    "bootstrap": "^4.5.2",
+    "@angular/animations": "^15.0.1",
+    "@angular/cdk": "^15.0.0",
+    "@angular/common": "^15.0.1",
+    "@angular/compiler": "^15.0.1",
+    "@angular/core": "^15.0.1",
+    "@angular/forms": "^15.0.1",
+    "@angular/platform-browser": "^15.0.1",
+    "@angular/platform-browser-dynamic": "^15.0.1",
+    "@angular/platform-server": "^15.0.1",
+    "@angular/router": "^15.0.1",
+    "@angular/service-worker": "^15.0.1",
+    "@apollo/client": "^3.7.1",
+    "@fortawesome/angular-fontawesome": "^0.12.0",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-brands-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@nguniversal/express-engine": "^15.0.0",
+    "@tailwindcss/forms": "^0.5.3",
+    "apollo-angular": "^4.1.1",
     "compression": "^1.7.4",
     "core-js": "^3.6.5",
     "express": "^4.15.2",
-    "graphql": "^16.5.0",
-    "photoswipe": "^4.1.3",
+    "graphql": "^16.6.0",
+    "photoswipe": "^5.3.4",
     "rxjs": "~6.6.3",
     "tslib": "^2.4.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.6",
-    "@angular/cli": "^13.3.6",
-    "@angular/compiler-cli": "^13.3.9",
-    "@angular/language-service": "^13.3.9",
-    "@graphql-codegen/add": "^2.0.1",
-    "@graphql-codegen/cli": "^1.13.2",
-    "@graphql-codegen/fragment-matcher": "^1.17.8",
-    "@graphql-codegen/typescript": "^1.13.2",
-    "@graphql-codegen/typescript-compatibility": "^2.0.0",
-    "@graphql-codegen/typescript-operations": "^1.13.2",
-    "@nguniversal/builders": "^13.1.1",
+    "@angular-devkit/build-angular": "^15.0.1",
+    "@angular/cli": "^15.0.1",
+    "@angular/compiler-cli": "^15.0.1",
+    "@angular/language-service": "^15.0.1",
+    "@graphql-codegen/add": "^3.2.1",
+    "@graphql-codegen/cli": "^2.14.0",
+    "@graphql-codegen/fragment-matcher": "^3.3.1",
+    "@graphql-codegen/typescript": "^2.8.2",
+    "@graphql-codegen/typescript-operations": "^2.5.7",
+    "@nguniversal/builders": "^15.0.0",
     "@types/compression": "^1.7.0",
     "@types/express": "^4.17.8",
     "@types/graphql": "^14.0.5",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "^15.3.1",
-    "@types/photoswipe": "^4.1.0",
+    "autoprefixer": "^10.4.13",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -73,11 +72,13 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^15.0.1",
+    "postcss": "^8.4.18",
+    "tailwindcss": "^3.2.1",
     "ts-loader": "^8.0.4",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.6.4",
+    "typescript": "~4.8.4",
     "webpack-bundle-analyzer": "^3.9.0",
     "webpack-cli": "^3.3.12"
   }
