@@ -1,8 +1,13 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { filter, map } from 'rxjs/operators';
 
-import { GetActiveCustomer, UpdateCustomerDetails, UpdateCustomerInput } from '../../../common/generated-types';
+import {
+    GetActiveCustomerQuery,
+    UpdateCustomerDetailsMutation,
+    UpdateCustomerDetailsMutationVariables,
+    UpdateCustomerInput
+} from '../../../common/generated-types';
 import { GET_ACTIVE_CUSTOMER } from '../../../common/graphql/documents.graphql';
 import { notNullOrUndefined } from '../../../common/utils/not-null-or-undefined';
 import { DataService } from '../../../core/providers/data/data.service';
@@ -12,18 +17,20 @@ import { UPDATE_CUSTOMER_DETAILS } from './account-customer-details.graphql';
 @Component({
     selector: 'vsf-account-customer-details',
     templateUrl: './account-customer-details.component.html',
-    styleUrls: ['./account-customer-details.component.scss'],
+    // styleUrls: ['./account-customer-details.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AccountCustomerDetailsComponent implements OnInit {
 
-    form: FormGroup;
+    form: UntypedFormGroup;
+
     constructor(private dataService: DataService,
-                private formBuilder: FormBuilder,
-                private changeDetectorRef: ChangeDetectorRef) { }
+                private formBuilder: UntypedFormBuilder,
+                private changeDetectorRef: ChangeDetectorRef) {
+    }
 
     ngOnInit() {
-        this.dataService.query<GetActiveCustomer.Query>(GET_ACTIVE_CUSTOMER, {}, 'network-only').pipe(
+        this.dataService.query<GetActiveCustomerQuery>(GET_ACTIVE_CUSTOMER, {}, 'network-only').pipe(
             map(data => data.activeCustomer),
             filter(notNullOrUndefined),
         ).subscribe(customer => {
@@ -43,7 +50,7 @@ export class AccountCustomerDetailsComponent implements OnInit {
             lastName: formValue.lastName,
             phoneNumber: formValue.phoneNumber,
         };
-        this.dataService.mutate<UpdateCustomerDetails.Mutation, UpdateCustomerDetails.Variables>(UPDATE_CUSTOMER_DETAILS, {
+        this.dataService.mutate<UpdateCustomerDetailsMutation, UpdateCustomerDetailsMutationVariables>(UPDATE_CUSTOMER_DETAILS, {
             input,
         }).subscribe(() => {
             this.form.markAsPristine();
