@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Input, OnDestroy, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
@@ -7,14 +7,14 @@ import { debounceTime } from 'rxjs/operators';
 @Component({
     selector: 'vsf-product-search-bar',
     templateUrl: './product-search-bar.component.html',
-    styleUrls: ['./product-search-bar.component.scss'],
+    // styleUrls: ['./product-search-bar.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductSearchBarComponent implements OnInit, OnDestroy {
     /** If true, searches as you type */
     @Input() autoSearch = false;
 
-    searchTerm = new FormControl('');
+    searchTerm = new UntypedFormControl('');
     private subscription: Subscription;
     constructor(private route: ActivatedRoute, private router: Router) { }
 
