@@ -3,7 +3,7 @@ import { ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs';
 import { filter, map, switchMap } from 'rxjs/operators';
 
-import { GetOrder } from '../../../common/generated-types';
+import { GetOrderQuery, GetOrderQueryVariables } from '../../../common/generated-types';
 import { notNullOrUndefined } from '../../../common/utils/not-null-or-undefined';
 import { DataService } from '../../../core/providers/data/data.service';
 
@@ -17,7 +17,7 @@ import { GET_ORDER } from './account-order-detail.graphql';
 })
 export class AccountOrderDetailComponent implements OnInit {
 
-    order$: Observable<GetOrder.OrderByCode | undefined>;
+    order$: Observable<GetOrderQuery['orderByCode'] | undefined>;
     constructor(private dataService: DataService, private route: ActivatedRoute) { }
 
     ngOnInit() {
@@ -25,7 +25,7 @@ export class AccountOrderDetailComponent implements OnInit {
             map(pm => pm.get('code')),
             filter(notNullOrUndefined),
             switchMap(code => {
-                return this.dataService.query<GetOrder.Query, GetOrder.Variables>(GET_ORDER, { code });
+                return this.dataService.query<GetOrderQuery, GetOrderQueryVariables>(GET_ORDER, { code });
             }),
             map(data => data.orderByCode),
         );
