@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { RequestPasswordReset } from '../../../common/generated-types';
+import { RequestPasswordResetMutation, RequestPasswordResetMutationVariables } from '../../../common/generated-types';
 import { DataService } from '../../../core/providers/data/data.service';
 
 import { REQUEST_PASSWORD_RESET } from './forgotten-password.graphql';
@@ -15,12 +15,13 @@ import { REQUEST_PASSWORD_RESET } from './forgotten-password.graphql';
 export class ForgottenPasswordComponent {
     emailAddress = '';
     submitted = false;
+
     constructor(private dataService: DataService, private route: ActivatedRoute) {
-        this.emailAddress = this.route.snapshot.paramMap.get('email') || '';
+        this.emailAddress = this.route.snapshot.paramMap.get('email') ?? '';
     }
 
     resetPassword() {
-        this.dataService.mutate<RequestPasswordReset.Mutation, RequestPasswordReset.Variables>(REQUEST_PASSWORD_RESET, {
+        this.dataService.mutate<RequestPasswordResetMutation, RequestPasswordResetMutationVariables>(REQUEST_PASSWORD_RESET, {
             emailAddress: this.emailAddress,
         }).subscribe(() => {
             this.submitted = true;
