@@ -1,9 +1,9 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
-  selector: 'vsf-checkout-sign-in',
-  templateUrl: './checkout-sign-in.component.html',
-  styleUrls: ['./checkout-sign-in.component.scss'],
+    selector: 'vsf-checkout-sign-in',
+    templateUrl: './checkout-sign-in.component.html',
+    styleUrls: ['./checkout-sign-in.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CheckoutSignInComponent {
