@@ -2,26 +2,25 @@ import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map, switchMap, take } from 'rxjs/operators';
 
+import { GetActiveCustomerQuery } from '../../../common/generated-types';
+import { GET_ACTIVE_CUSTOMER } from '../../../common/graphql/documents.graphql';
 import { DataService } from '../../providers/data/data.service';
 import { StateService } from '../../providers/state/state.service';
-
-import { GetActiveCustomer } from '../../../common/generated-types';
-import { GET_ACTIVE_CUSTOMER } from '../../../common/graphql/documents.graphql';
 
 @Component({
     selector: 'vsf-account-link',
     templateUrl: './account-link.component.html',
-    styleUrls: ['./account-link.component.scss'],
+    // styleUrls: ['./account-link.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AccountLinkComponent implements OnInit {
 
-    activeCustomer$: Observable<GetActiveCustomer.ActiveCustomer | undefined>;
+    activeCustomer$: Observable<GetActiveCustomerQuery['activeCustomer']>;
     constructor(private dataService: DataService,
                 private stateService: StateService) {}
 
     ngOnInit() {
-        const getActiveCustomer$ = this.dataService.query<GetActiveCustomer.Query>(GET_ACTIVE_CUSTOMER, {}, 'network-only');
+        const getActiveCustomer$ = this.dataService.query<GetActiveCustomerQuery>(GET_ACTIVE_CUSTOMER, {}, 'network-only');
 
         getActiveCustomer$.pipe(take(1)).subscribe(data => {
             if (data.activeCustomer) {
@@ -35,7 +34,7 @@ export class AccountLinkComponent implements OnInit {
         );
     }
 
-    userName(customer: NonNullable<GetActiveCustomer.ActiveCustomer>): string {
+    userName(customer: NonNullable<GetActiveCustomerQuery['activeCustomer']>): string {
         const { firstName, lastName, emailAddress } = customer;
         if (firstName && lastName) {
             return `${firstName} ${lastName}`;
