@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core';
 
-import { GetCollection } from '../../../common/generated-types';
+import { GetCollectionQuery } from '../../../common/generated-types';
 
 @Component({
     selector: 'vsf-collection-breadcrumbs',
@@ -10,7 +10,7 @@ import { GetCollection } from '../../../common/generated-types';
 })
 export class CollectionBreadcrumbsComponent {
 
-    @Input() breadcrumbs: GetCollection.Breadcrumbs[] = [];
+    @Input() breadcrumbs?: NonNullable<GetCollectionQuery['collection']>['breadcrumbs'] = [];
     @Input() linkLast = false;
 
     tail<T>(arr: T[] | null): T[] {
