@@ -47,7 +47,7 @@ export class LayoutHeaderComponent implements AfterViewInit, OnDestroy {
     }
 
     private setUpScrollHandler(_window: Window) {
-        this.subscription = fromEvent(_window, 'scroll').pipe(
+        this.subscription = fromEvent(_window, 'scroll', { passive: true }).pipe(
             map(() => _window.scrollY),
             bufferTime(250),
             filter(val => 1 < val.length),
