@@ -3,7 +3,7 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 @Component({
     selector: 'vsf-layout',
     templateUrl: './layout.component.html',
-    styleUrls: ['./layout.component.scss'],
+    // styleUrls: ['./layout.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LayoutComponent {}
