@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { SearchProducts } from '../../../common/generated-types';
+import { SearchProductsQuery } from '../../../common/generated-types';
 import { getRouteArrayParam } from '../../../common/utils/get-route-array-param';
 
 export interface FacetWithValues {
@@ -17,12 +17,12 @@ export interface FacetWithValues {
 @Component({
     selector: 'vsf-product-list-controls',
     templateUrl: './product-list-controls.component.html',
-    styleUrls: ['./product-list-controls.component.scss'],
+    // styleUrls: ['./product-list-controls.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductListControlsComponent implements OnChanges {
     @Input() activeFacetValueIds: string[] = [];
-    @Input() facetValues: SearchProducts.FacetValues[] | null;
+    @Input() facetValues: SearchProductsQuery['search']['facetValues'] | null;
     @Input() totalResults = 0;
     facets: FacetWithValues[];
     manuallyExpanded = false;
@@ -65,7 +65,7 @@ export class ProductListControlsComponent implements OnChanges {
         return item.id;
     }
 
-    private groupFacetValues(facetValues: SearchProducts.FacetValues[] | null): FacetWithValues[] {
+    private groupFacetValues(facetValues: SearchProductsQuery['search']['facetValues'] | null): FacetWithValues[] {
         if (!facetValues) {
             return [];
         }
