@@ -1,16 +1,16 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 
-import { Address, OrderAddress } from '../../../common/generated-types';
+import { AddressFragment, OrderAddressFragment } from '../../../common/generated-types';
 import { DataService } from '../../../core/providers/data/data.service';
 
 @Component({
     selector: 'vsf-address-card',
     templateUrl: './address-card.component.html',
-    styleUrls: ['./address-card.component.scss'],
+    // styleUrls: ['./address-card.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AddressCardComponent {
-    @Input() address: OrderAddress.Fragment | Address.Fragment;
+    @Input() address: OrderAddressFragment | AddressFragment;
     @Input() title = '';
 
     getCountryName(): string {
