@@ -1,17 +1,17 @@
 import { ChangeDetectionStrategy, Component, Input, OnChanges } from '@angular/core';
 import { DomSanitizer, SafeStyle } from '@angular/platform-browser';
 
-import { GetCollection } from '../../../common/generated-types';
+import { GetCollectionQuery } from '../../../common/generated-types';
 import { AssetPreviewPipe } from '../../pipes/asset-preview.pipe';
 
 @Component({
     selector: 'vsf-collection-card',
     templateUrl: './collection-card.component.html',
-    styleUrls: ['./collection-card.component.scss'],
+    // styleUrls: ['./collection-card.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CollectionCardComponent implements OnChanges {
-    @Input() collection: GetCollection.Children;
+    @Input() collection: NonNullable<GetCollectionQuery['collection']>;
     backgroundImage: SafeStyle;
 
     constructor(private sanitizer: DomSanitizer) {}
