@@ -1,12 +1,12 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { Asset } from '../../common/generated-types';
+import { AssetFragment } from '../../common/generated-types';
 
 @Pipe({
     name: 'assetPreview',
 })
 export class AssetPreviewPipe implements PipeTransform {
-    transform(asset?: Asset.Fragment, ...args: Array<string | number>): string {
+    transform(asset?: AssetFragment, ...args: Array<string | number>): string {
         if (!asset) {
             return '';
         }
@@ -16,7 +16,7 @@ export class AssetPreviewPipe implements PipeTransform {
         const previewUrl = asset.preview.replace(/\\/g, '/');
         const fp = asset.focalPoint ? `&fpx=${asset.focalPoint.x}&fpy=${asset.focalPoint.y}` : '';
         const query = this.getSizeQuery(args);
-        return `${previewUrl}?${query}${fp}`;
+        return `${previewUrl}?${query}${fp}&format=webp`;
     }
 
     private getSizeQuery(args?: Array<string | number>): string {
