@@ -8,4 +8,4 @@ if (environment.production) {
 
 export { AppServerModule } from './app/app.server.module';
 
-export { renderModule, renderModuleFactory } from '@angular/platform-server';
+export { renderModuleFactory } from '@angular/platform-server';
